RANGER-2927: added packaging for Python client
(cherry picked from commit 758ef30008d4f4da430417f8d69285679444a562)
