+++ intg/src/main/python/apache_ranger/model/ranger_base.py
