+++ ranger-examples/sample-client/src/main/resources/logging.conf
